gateware: Create ice40 module.
Contains;

 * UP5K's SPRAM
 * LED hard IP.
