+++ targets/icebreaker/base.py
@@ -11,7 +11,7 @@ from litex.soc.integration.soc_core import *
 from litex.soc.integration.builder import *
 from litex.soc.interconnect.csr import AutoCSR, CSRStatus, CSRStorage
 
-from gateware import up5kspram
+from gateware import ice40
 from gateware import cas
 from gateware import spi_flash
 
@@ -57,55 +57,6 @@ class _CRG(Module):
             )
         self.specials += AsyncResetSynchronizer(self.cd_por, reset)
 
-
-class _SBLED(Module, AutoCSR):
-    def __init__(self, pads):
-
-        rgba_pwm = Signal(3)
-
-        self.dat = CSRStorage(8)
-        self.addr = CSRStorage(4)
-        self.ctrl = CSRStorage(4)
-
-        self.specials += Instance("SB_RGBA_DRV",
-            i_CURREN = self.ctrl.storage[1],
-            i_RGBLEDEN = self.ctrl.storage[2],
-            i_RGB0PWM = rgba_pwm[0],
-            i_RGB1PWM = rgba_pwm[1],
-            i_RGB2PWM = rgba_pwm[2],
-            o_RGB0 = pads.rgb0,
-            o_RGB1 = pads.rgb1,
-            o_RGB2 = pads.rgb2,
-            p_CURRENT_MODE = "0b1",
-            p_RGB0_CURRENT = "0b000011",
-            p_RGB1_CURRENT = "0b000001",
-            p_RGB2_CURRENT = "0b000011",
-        )
-
-        self.specials += Instance("SB_LEDDA_IP",
-            i_LEDDCS = self.dat.re,
-            i_LEDDCLK = ClockSignal(),
-            i_LEDDDAT7 = self.dat.storage[7],
-            i_LEDDDAT6 = self.dat.storage[6],
-            i_LEDDDAT5 = self.dat.storage[5],
-            i_LEDDDAT4 = self.dat.storage[4],
-            i_LEDDDAT3 = self.dat.storage[3],
-            i_LEDDDAT2 = self.dat.storage[2],
-            i_LEDDDAT1 = self.dat.storage[1],
-            i_LEDDDAT0 = self.dat.storage[0],
-            i_LEDDADDR3 = self.addr.storage[3],
-            i_LEDDADDR2 = self.addr.storage[2],
-            i_LEDDADDR1 = self.addr.storage[1],
-            i_LEDDADDR0 = self.addr.storage[0],
-            i_LEDDDEN = self.dat.re,
-            i_LEDDEXE = self.ctrl.storage[0],
-            # o_LEDDON = led_is_on, # Indicates whether LED is on or not
-            # i_LEDDRST = ResetSignal(), # This port doesn't actually exist
-            o_PWMOUT0 = rgba_pwm[0], 
-            o_PWMOUT1 = rgba_pwm[1], 
-            o_PWMOUT2 = rgba_pwm[2],
-            o_LEDDON = Signal(), 
-        )
 
 class BaseSoC(SoCCore):
     csr_peripherals = (
@@ -156,7 +107,7 @@ class BaseSoC(SoCCore):
 
         # rgb led connector
         platform.add_extension(icebreaker.rgb_led)
-        self.submodules.rgbled = _SBLED(platform.request("rgbled", 0))
+        self.submodules.rgbled = ice40.LED(platform.request("rgbled", 0))
         self.add_csr("rgbled")
 
         bios_size = 0x8000
@@ -166,7 +117,7 @@ class BaseSoC(SoCCore):
 
         # SPRAM- UP5K has single port RAM, might as well use it as SRAM to
         # free up scarce block RAM.
-        self.submodules.spram = up5kspram.Up5kSPRAM(size=128*1024)
+        self.submodules.spram = ice40.SPRAM(size=128*1024)
         self.register_mem("sram", 0x10000000, self.spram.bus, 0x20000)
 
         # We don't have a DRAM, so use the remaining SPI flash for user
