+++ targets/upduino_v1/base.py
@@ -7,17 +7,17 @@ from migen import *
 from migen.genlib.resetsync import AsyncResetSynchronizer
 from migen.fhdl import verilog
 
-
 from litex.build.generic_platform import Pins, Subsignal, IOStandard
 from litex.soc.integration.soc_core import *
 from litex.soc.integration.builder import *
 
-from gateware import up5kspram
 from gateware import cas
+from gateware import ice40
 from gateware import spi_flash
 
 from targets.utils import csr_map_update
 import platforms.upduino_v1 as upduino
+
 
 class _CRG(Module):
     def __init__(self, platform):
@@ -91,7 +91,7 @@ class BaseSoC(SoCCore):
 
         # SPRAM- UP5K has single port RAM, might as well use it as SRAM to
         # free up scarce block RAM.
-        self.submodules.spram = up5kspram.Up5kSPRAM(size=128*1024)
+        self.submodules.spram = ice40.SPRAM(size=128*1024)
         self.register_mem("sram", 0x10000000, self.spram.bus, 0x20000)
 
         # We don't have a DRAM, so use the remaining SPI flash for user
@@ -106,5 +106,6 @@ class BaseSoC(SoCCore):
         # onto softcore's address bus.
         platform.toolchain.build_template[3] = "icepack -s {build_name}.txt {build_name}.bin"
         platform.toolchain.nextpnr_build_template[2] = "icepack -s {build_name}.txt {build_name}.bin"
+
 
 SoC = BaseSoC
